Updating flow for building assets

+++ buildAllCosmereBundles.ps1
@@ -1,18 +1,38 @@
-if ($IsWindows) {
+if ($IsWindows)
+{
     $unityPath = "C:\Program Files\Unity\Hub\Editor\2022.3.35f1\Editor\Unity.exe"
-} elseif ($IsMacOS) {
-    $unityPath = "/Applications/Unity/Hub/Editor/2022.3.35f1/Unity.app/Contents/MacOS/Unity"
-} else {
-    throw "Unsupported platform. Only Windows and macOS are supported."
+    $buildTarget = "windows"
 }
+elseif ($IsMacOS)
+{
+    $unityPath = "/Applications/Unity/Hub/Editor/2022.3.35f1/Unity.app/Contents/MacOS/Unity"
+    $buildTarget = "mac"
+}
+else
+{
+    $unityPath = $Env:UNITY_PATH
+    $buildTarget = $Env:UNITY_BUILD_TARGET;
+}
+
+if ( [string]::IsNullOrEmpty($unityPath))
+{
+    Write-Host "Could not find unityPath. If you are on Windows or Mac, make sure your powershell version is up to date (v7)"
+    Write-Host "If you are not, or don't want to update, set your UNITY_PATH environment variable"
+    exit
+}
+
+if ( [string]::IsNullOrEmpty($buildTarget))
+{
+    Write-Host "Could not find buildTarget. If you are on Windows or Mac, make sure your powershell version is up to date (v7)"
+    Write-Host "If you are not, or don't want to update, set your UNITY_BUILD_TARGET environment variable (windows, mac, linux)"
+    exit
+}
+
 
 $cleanPath = "..\AssetBuilder\Assets\Data"
 $assetOutput = "..\AssetBuilder\Assets\AssetBundles"
 Write-Host "Cleaning $cleanPath"
 Remove-Item -Recurse -Force -ErrorAction SilentlyContinue "$cleanPath\*" | Out-Null
-
-# List of modules to process
-$mods = @("Core", "Resources", "Framework", "Scadrial")
 
 function Get-FolderHash($folderPath)
 {
@@ -39,16 +59,18 @@ function Get-FolderHash($folderPath)
 }
 
 
+# List of modules to process
+$mods = Get-ChildItem -Directory -Name |
+        Where-Object { $_ -like 'Cosmere*' } |
+        ForEach-Object { $_ -replace '^Cosmere', '' }
 foreach ($mod in $mods)
 {
     Write-Host "--------------------------------------"
     Write-Host "Processing Cosmere$mod..."
 
-    $modLower = $mod.ToLower()
     $srcAssets = "..\RimworldCosmere\Cosmere$mod\Assets"
     $bundleName = "CryptikLemur.Cosmere.$mod"
     $destPath = "..\AssetBuilder\Assets\Data\$bundleName"
-    $bundleOutput = "resource_cosmere_$modLower"
     $finalOutput = "..\RimworldCosmere\Cosmere$mod\AssetBundles"
     $hashFile = "..\RimworldCosmere\Cosmere$mod\.lastassetbuildhash"
 
@@ -81,57 +103,22 @@ foreach ($mod in $mods)
             "-quit",
             '-projectPath="..\AssetBuilder"',
             "-executeMethod", "ModAssetBundleBuilder.BuildBundles",
-            "--assetBundleName=$bundleName"
+            "--assetBundleName=$bundleName",
+            "--buildTarget=$buildTarget",
+            "--outputLocation=$finalOutput"
         )
 
         Write-Host "    Building asset bundle: $bundleName"
-        $process = Start-Process -FilePath $unityPath -ArgumentList $unityArgs -Wait -PassThru
+        $process = Start-Process $unityPath -ArgumentList $unityArgs -Wait -PassThru
 
         if ($process.ExitCode -ne 0)
         {
             Write-Host "    Unity failed for $mod (exit code $( $process.ExitCode )). Skipping bundle move."
-            exit
-            Remove-Item -Recurse -Force -ErrorAction SilentlyContinue $destPath | Out-Null
             continue
-        }
-
-        if (Test-Path $finalOutput)
-        {
-            Write-Host "    Cleaning $finalOutput"
-            Remove-Item -Recurse -Force -ErrorAction SilentlyContinue "$finalOutput\*" | Out-Null
-        }
-        else
-        {
-            New-Item -ItemType Directory -Path $finalOutput | Out-Null
-        }
-
-        $platforms = @("windows", "mac", "linux")
-        $foundAny = $false
-        foreach ($platform in $platforms) {
-            $platformOutput = "$assetOutput\$platform\$bundleName_$platform"
-            $platformManifest = "$platformOutput.manifest"
-
-            if (Test-Path $platformOutput) {
-                Move-Item $platformOutput "$finalOutput/${bundleOutput}_$platform" -Force
-                $foundAny = $true
-            } else {
-                Write-Host "    Warning: $platformOutput not found."
-            }
-
-            if (Test-Path $platformManifest) {
-                Move-Item $platformManifest "$finalOutput/${bundleOutput}_$platform.manifest" -Force
-            } else {
-                Write-Host "    Warning: $platformManifest not found."
-            }
-        }
-
-        if (-not $foundAny) {
-            Write-Host "    Warning: No asset bundles found for any platform."
         }
 
         $currentHash | Out-File -Encoding ASCII -FilePath $hashFile
         Write-Host "    Done with Cosmere$mod."
-        Remove-Item -Recurse -Force -ErrorAction SilentlyContinue $destPath | Out-Null
     }
     else
     {
